functions/social_networks/facebook/post_comment/requirements.txt
requests==2.25.1
pysocks
stem==1.8.0
tenacity==6.3.1
marshmallow==3.10.0
requests_toolbelt
dateparser==1.0.0